pyproject.toml
[project]
name = "voice2message-bot"
version = "3.0.0"
description = "Telegram bot that transcribes voice messages using ElevenLabs Scribe v2"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.3",
    "elevenlabs>=1.0.0",
    "openai>=1.35.0",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
